Fixed various string format specifier issues. Patch by Jan Weiß.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/CoreObject@7951 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ COArrayDiff.mm
@@ -166,7 +166,7 @@ public:
 
 - (NSString *)description
 {
-	return [NSString stringWithFormat: @"Insert '%@' at %d", insertedObjects, range.location];
+	return [NSString stringWithFormat: @"Insert '%@' at %lu", insertedObjects, (unsigned long)range.location];
 }
 
 - (BOOL) isEqual: (id)other
